move all query related methods from RODSConn to RC_query trait
